requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
Pillow
numpy
opencv-python
python-multipart
pytesseract
sympy